Fixes the sometimes missing cover in WorkLayout.
The pids are used as key in `CoverImage.store`, and if `this.props.work.pid`
changes, then the lookup of the cover in the store will fail.
(`this.props.work` will change when the metadata is loaded).

The material pid in `this.props.id` should be sufficient for the cover
service, and thus removing `this.props.work.pid` from the list of pids
is ok, and solves the problem.